Use helm chart to install Kratix
There are a lot of comments in this PR which will be
removed when the WIP PR that is open makes this a bit more
usable on initial install (#8).

Also note, that the Namespace is `kratix-platform-system`
and not `kratix`. I am not sure if you expect namespace
and name of folder to be the same so wanted to verify.