feat(cluster-autoscaler): add cluster autoscaler for aws
Note for when I'm back. This is the working autoscaler stuff, but the
permissions come from the node. It doesn't have the correct IAM permissions
for this so we need to figure out how to add that. I don't really want
to do that in the KF installation, but ideally via a Terraform job which
we can run in here. Annoyingly, I've lost access to AWS so I can't check
my hypothesis out until I get that back.